requirements.txt
python-binance==1.0.19
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
prompt_toolkit==3.0.43
